+++ library/phe/src/vsce_phe_cipher.c
@@ -366,7 +366,31 @@ VSCE_PUBLIC vsce_status_t
 vsce_phe_cipher_encrypt(
         vsce_phe_cipher_t *self, vsc_data_t plain_text, vsc_data_t account_key, vsc_buffer_t *cipher_text) {
 
+    return vsce_phe_cipher_auth_encrypt(self, plain_text, vsc_data_empty(), account_key, cipher_text);
+}
+
+//
+//  Decrypts data using account key
+//
+VSCE_PUBLIC vsce_status_t
+vsce_phe_cipher_decrypt(
+        vsce_phe_cipher_t *self, vsc_data_t cipher_text, vsc_data_t account_key, vsc_buffer_t *plain_text) {
+
+    return vsce_phe_cipher_auth_decrypt(self, cipher_text, vsc_data_empty(), account_key, plain_text);
+}
+
+//
+//  Encrypts data (and authenticates additional data) using account key
+//
+VSCE_PUBLIC vsce_status_t
+vsce_phe_cipher_auth_encrypt(vsce_phe_cipher_t *self, vsc_data_t plain_text, vsc_data_t additional_data,
+        vsc_data_t account_key, vsc_buffer_t *cipher_text) {
+
     VSCE_ASSERT_PTR(self);
+    VSCE_ASSERT(vsc_data_is_valid(plain_text));
+    VSCE_ASSERT(vsc_data_is_valid(additional_data));
+    VSCE_ASSERT(vsc_data_is_valid(account_key));
+    VSCE_ASSERT(vsc_buffer_is_valid(cipher_text));
     VSCE_ASSERT(account_key.len == vsce_phe_common_PHE_ACCOUNT_KEY_LENGTH);
     VSCE_ASSERT(plain_text.len <= vsce_phe_common_PHE_MAX_ENCRYPT_LEN);
     VSCE_ASSERT(vsc_buffer_capacity(cipher_text) >= vsce_phe_cipher_encrypt_len(self, plain_text.len));
@@ -410,7 +434,7 @@ vsce_phe_cipher_encrypt(
     memcpy(vsc_buffer_unused_bytes(cipher_text), salt, sizeof(salt));
     vsc_buffer_inc_used(cipher_text, sizeof(salt));
 
-    f_status = vscf_aes256_gcm_encrypt(aes256_gcm, plain_text, cipher_text);
+    f_status = vscf_aes256_gcm_auth_encrypt(aes256_gcm, plain_text, additional_data, cipher_text, NULL);
 
     if (f_status != vscf_status_SUCCESS) {
         status = vsce_status_ERROR_AES_FAILED;
@@ -429,13 +453,17 @@ rng_err:
 }
 
 //
-//  Decrypts data using account key
+//  Decrypts data (and verifies additional data) using account key
 //
 VSCE_PUBLIC vsce_status_t
-vsce_phe_cipher_decrypt(
-        vsce_phe_cipher_t *self, vsc_data_t cipher_text, vsc_data_t account_key, vsc_buffer_t *plain_text) {
+vsce_phe_cipher_auth_decrypt(vsce_phe_cipher_t *self, vsc_data_t cipher_text, vsc_data_t additional_data,
+        vsc_data_t account_key, vsc_buffer_t *plain_text) {
 
     VSCE_ASSERT_PTR(self);
+    VSCE_ASSERT(vsc_data_is_valid(cipher_text));
+    VSCE_ASSERT(vsc_data_is_valid(additional_data));
+    VSCE_ASSERT(vsc_data_is_valid(account_key));
+    VSCE_ASSERT(vsc_buffer_is_valid(plain_text));
     VSCE_ASSERT(account_key.len == vsce_phe_common_PHE_ACCOUNT_KEY_LENGTH);
     VSCE_ASSERT(cipher_text.len <= vsce_phe_common_PHE_MAX_DECRYPT_LEN);
     VSCE_ASSERT(vsc_buffer_capacity(plain_text) >= vsce_phe_cipher_decrypt_len(self, cipher_text.len));
@@ -463,9 +491,9 @@ vsce_phe_cipher_decrypt(
     vscf_aes256_gcm_set_nonce(
             aes256_gcm, vsc_data_slice_end(vsc_buffer_data(&derived_secret_buf), 0, vsce_phe_cipher_NONCE_LEN));
 
-    vscf_status_t f_status = vscf_aes256_gcm_decrypt(aes256_gcm,
+    vscf_status_t f_status = vscf_aes256_gcm_auth_decrypt(aes256_gcm,
             vsc_data_slice_beg(cipher_text, vsce_phe_cipher_SALT_LEN, cipher_text.len - vsce_phe_cipher_SALT_LEN),
-            plain_text);
+            additional_data, vsc_data_empty(), plain_text);
 
     if (f_status != vscf_status_SUCCESS) {
         status = vsce_status_ERROR_AES_FAILED;